requirements.txt
fastapi
uvicorn
python-dotenv
openai
pika
dependency-injector
boto3
aiohttp